moving QueryOutputHandler to NetworkConfig dir
